requirements.txt
asgiref==3.8.1
blinker==1.8.2
channels==4.1.0
chardet==5.2.0
click==8.1.7
colorama==0.4.6
dash==2.9.3
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==4.2.18
django-modeltranslation==0.19.4
django-plotly-dash==2.3.1
dpd_components==0.1.0
et-xmlfile==1.1.0
filetype==1.2.0
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
pyClamd==0.4.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2024.1
reportlab==4.2.5
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tenacity==8.3.0
tzdata==2024.1
Werkzeug==3.0.3
XlsxWriter==3.2.2
whitenoise
Django==4.2
gunicorn
whitenoise
weasyprint
